students/k3339/Zoteev_Maksim/Lr1/lab/requirements.txt
fastapi[all]
sqlmodel
psycopg2-binary
alembic
python-dotenv
pydantic-settings
PyJWT
passlib[bcrypt]
